✨ Rename sort.py to authentication.py
